refactor: feedImageInfo 관련 불필요한 필드 제거 및 구조 간소화 (#161)
- SaveFeedRequestDto 및 ModifyFeedRequestDto에서 feedImageInfos 필드 제거
- feedImageOrders의 타입을 List<feedImageOrder>에서 List<Integer>로 단순화
- 코드 구조 간소화 및 사용하지 않는 필드 제거

+++ service/src/main/java/way/application/service/hideFeed/dto/response/HideFeedResponseDto.java
@@ -26,7 +26,7 @@ public class HideFeedResponseDto {
 	public record hideFeedImageInfo(
 		Long feedImageSeq,
 		String feedImageURL,
-		Long feedImageOrder
+		Integer feedImageOrder
 	) {
 
 	}
